Trading_Process/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Google Cloud
google-cloud-bigquery>=3.10.0
google-cloud-core>=2.3.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
pandas-gbq>=0.19.0

# AWS
boto3>=1.26.0
botocore>=1.29.0

# HTTP requests
requests>=2.28.0
urllib3>=1.26.0

# Environment variables
python-dotenv>=1.0.0

# Progress bars
tqdm>=4.64.0

# Date/time
python-dateutil>=2.8.0
pytz>=2022.7

# Email and SMTP
# (using built-in smtplib, no additional packages needed)

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.0.0 